+++ api/types/second_factor.go
@@ -1,0 +1,172 @@
+/*
+Copyright 2024 Gravitational, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package types
+
+import (
+	"encoding/json"
+	"slices"
+
+	"github.com/gravitational/trace"
+
+	"github.com/gravitational/teleport/api/constants"
+)
+
+// legacySecondFactorFromSecondFactors returns a suitable legacy second factor for the given list of second factors.
+func legacySecondFactorFromSecondFactors(secondFactors []SecondFactorType) constants.SecondFactorType {
+	hasOTP := slices.Contains(secondFactors, SecondFactorType_SECOND_FACTOR_TYPE_OTP)
+	hasWebAuthn := slices.Contains(secondFactors, SecondFactorType_SECOND_FACTOR_TYPE_WEBAUTHN)
+
+	switch {
+	case hasOTP && hasWebAuthn:
+		return constants.SecondFactorOn
+	case hasWebAuthn:
+		return constants.SecondFactorWebauthn
+	case hasOTP:
+		return constants.SecondFactorOTP
+	default:
+		return constants.SecondFactorOff
+	}
+}
+
+// secondFactorsFromLegacySecondFactor returns the list of SecondFactorTypes supported by the given second factor type.
+func secondFactorsFromLegacySecondFactor(sf constants.SecondFactorType) []SecondFactorType {
+	switch sf {
+	case constants.SecondFactorOff:
+		return nil
+	case constants.SecondFactorOptional, constants.SecondFactorOn:
+		return []SecondFactorType{SecondFactorType_SECOND_FACTOR_TYPE_WEBAUTHN, SecondFactorType_SECOND_FACTOR_TYPE_OTP}
+	case constants.SecondFactorOTP:
+		return []SecondFactorType{SecondFactorType_SECOND_FACTOR_TYPE_OTP}
+	case constants.SecondFactorWebauthn:
+		return []SecondFactorType{SecondFactorType_SECOND_FACTOR_TYPE_WEBAUTHN}
+	default:
+		return nil
+	}
+}
+
+// MarshalJSON marshals SecondFactorType to string.
+func (s *SecondFactorType) MarshalYAML() (interface{}, error) {
+	val, err := s.encode()
+	if err != nil {
+		return nil, trace.Wrap(err)
+	}
+	return val, nil
+}
+
+// UnmarshalYAML supports parsing SecondFactorType from string.
+func (s *SecondFactorType) UnmarshalYAML(unmarshal func(interface{}) error) error {
+	var val interface{}
+	err := unmarshal(&val)
+	if err != nil {
+		return trace.Wrap(err)
+	}
+
+	err = s.decode(val)
+	return trace.Wrap(err)
+}
+
+// MarshalJSON marshals SecondFactorType to string.
+func (s *SecondFactorType) MarshalJSON() ([]byte, error) {
+	val, err := s.encode()
+	if err != nil {
+		return nil, trace.Wrap(err)
+	}
+	out, err := json.Marshal(val)
+	return out, trace.Wrap(err)
+}
+
+// UnmarshalJSON supports parsing SecondFactorType from string.
+func (s *SecondFactorType) UnmarshalJSON(data []byte) error {
+	var val interface{}
+	err := json.Unmarshal(data, &val)
+	if err != nil {
+		return trace.Wrap(err)
+	}
+
+	err = s.decode(val)
+	return trace.Wrap(err)
+}
+
+const (
+	// secondFactorTypeOTPString is the string representation of SecondFactorType_SECOND_FACTOR_TYPE_OTP
+	secondFactorTypeOTPString = "otp"
+	// secondFactorTypeWebauthnString is the string representation of SecondFactorType_SECOND_FACTOR_TYPE_WEBAUTHN
+	secondFactorTypeWebauthnString = "webauthn"
+	// secondFactorTypeSSOString is the string representation of SecondFactorType_SECOND_FACTOR_TYPE_SSO
+	secondFactorTypeSSOString = "sso"
+)
+
+// ToString returns the user friendly string representation of the second factor type.
+func (s *SecondFactorType) ToString() string {
+	str, _ := s.encode()
+	return str
+}
+
+func (s *SecondFactorType) encode() (string, error) {
+	switch *s {
+	case SecondFactorType_SECOND_FACTOR_TYPE_UNSPECIFIED:
+		return "", nil
+	case SecondFactorType_SECOND_FACTOR_TYPE_OTP:
+		return secondFactorTypeOTPString, nil
+	case SecondFactorType_SECOND_FACTOR_TYPE_WEBAUTHN:
+		return secondFactorTypeWebauthnString, nil
+	case SecondFactorType_SECOND_FACTOR_TYPE_SSO:
+		return secondFactorTypeSSOString, nil
+	default:
+		return "", trace.BadParameter("SecondFactorType invalid value %v", *s)
+	}
+}
+
+func (s *SecondFactorType) decode(val any) error {
+	switch v := val.(type) {
+	case string:
+		switch v {
+		case secondFactorTypeOTPString:
+			*s = SecondFactorType_SECOND_FACTOR_TYPE_OTP
+		case secondFactorTypeWebauthnString:
+			*s = SecondFactorType_SECOND_FACTOR_TYPE_WEBAUTHN
+		case secondFactorTypeSSOString:
+			*s = SecondFactorType_SECOND_FACTOR_TYPE_SSO
+		case "":
+			*s = SecondFactorType_SECOND_FACTOR_TYPE_UNSPECIFIED
+		default:
+			return trace.BadParameter("SecondFactorType invalid value %v", val)
+		}
+	case int32:
+		return trace.Wrap(s.setFromEnum(v))
+	case int64:
+		return trace.Wrap(s.setFromEnum(int32(v)))
+	case int:
+		return trace.Wrap(s.setFromEnum(int32(v)))
+	case float64:
+		return trace.Wrap(s.setFromEnum(int32(v)))
+	case float32:
+		return trace.Wrap(s.setFromEnum(int32(v)))
+	default:
+		return trace.BadParameter("RequireMFAType invalid type %T", val)
+	}
+	return nil
+}
+
+// setFromEnum sets the value from enum value as int32.
+func (r *SecondFactorType) setFromEnum(val int32) error {
+	if _, ok := SecondFactorType_name[val]; !ok {
+		return trace.BadParameter("invalid SecondFactorType mode %v", val)
+	}
+	*r = SecondFactorType(val)
+	return nil
+}
